Update 01-fizzbuzz.js
Initially I completed the homework on the main branch, so the files here don't show an actual change.

Completed fizzbuzz() function with a for loop containing an if/else statement for determining what to print to console for each value between 1-100.

Feedback request:  Anything regarding the code.  Anything regarding the formatting of this pull request.